+++ test/eca/llm_providers/anthropic_test.clj
@@ -7,13 +7,13 @@
 (deftest ->messages-with-history-test
   (testing "no previous history"
     (is (match?
-         [{:role "user" :content "Hey"}]
+         [{:role "user" :content "Hey" :cache_control {:type "ephemeral"}}]
          (#'llm-providers.anthropic/->messages-with-history [] "Hey"))))
   (testing "With basic text history"
     (is (match?
          [{:role "user" :content "Count with me: 1"}
           {:role "assistant" :content "2"}
-          {:role "user" :content "3"}]
+          {:role "user" :content "3" :cache_control {:type "ephemeral"}}]
          (#'llm-providers.anthropic/->messages-with-history
           [{:role "user" :content "Count with me: 1"}
            {:role "assistant" :content "2"}]
@@ -30,7 +30,7 @@
                                    :tool-use-id "call-1"
                                    :content "Allowed directories: /foo/bar\n"}]}
           {:role "assistant" :content "I see /foo/bar"}
-          {:role "user" :content "Thanks"}]
+          {:role "user" :content "Thanks" :cache_control {:type "ephemeral"}}]
          (#'llm-providers.anthropic/->messages-with-history
           [{:role "user" :content "List the files you are allowed"}
            {:role "assistant" :content "Ok!"}